Made logging fix neat like content_acceptance_tests, thanks etam!
